quosyntax: Initial fork of hclsyntax
The Quo language is intended to be a slight variant of HCL native syntax
that has some small adjustments to make it better suited to describing
financial operations.

It will still implement the HCL syntax-agnostic interfaces, and thus we
can think of Quo as being an alternative syntax for HCL, though one that
is not intended to interoperate with configuration written in other HCL
syntaxes.