quoty: Use cty's new CapsuleWithOps function
Rather than ending up re-implementing all of the basic operations like
RawEquals, conversions, etc, we'll instead let cty continue worrying about
those and just give it some hints on how to deal with them with our
special capsule types in particular.

This still needs a lot more work to get all of the quosyntax functionality
in order, because a lot of it is still expecting to deal with normal
cty.Number rather than quoty.Number, but at least we can now rely on the
type-generic cty functionality for our custom number type.